pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "monocle_apptrace"
version = "0.3.0b6"
authors = []
description = "package with monocle genAI tracing"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
   'requests',
   'wrapt>=1.14.0',
   'opentelemetry-api>=1.21.0',
   'opentelemetry-sdk>=1.21.0',
   'opentelemetry-instrumentation',
]

[tool.pytest.ini_options]
#log_cli = true
#log_cli_level = "INFO"
# Enable file logging
log_file = "traces.log"
log_file_level = "INFO"
log_format = "%(asctime)s %(levelname)s %(pathname)s:%(lineno)d %(message)s"
log_date_format = "%Y-%m-%d %H:%M:%S"

pythonpath = [
  "src",
  "tests"
]
markers = [
  "integration"
]

[project.optional-dependencies]
dev = [
  'langchain-openai==0.1.8',
  'langchain-chroma==0.1.1',
  'langchain-community==0.2.5',
  'langchain==0.2.5',
  'datasets==2.20.0',
  'numpy==1.26.4',
  'types-requests==2.31.0.20240106',
  'InstructorEmbedding==1.0.1',
  'sentence-transformers==2.6.1',
  'faiss-cpu==1.8.0',
  'pytest==8.0.0',
  'llama-index==0.10.30',
  'llama-index-embeddings-huggingface==0.2.0',
  'llama-index-vector-stores-chroma==0.1.9',
  'parameterized==0.9.0',
  'llama-index-llms-mistralai==0.1.20',
  'langchain-mistralai==0.1.13',
  'mistral-haystack==0.0.2',
  'langchain-aws==0.1.10',
  'azure-storage-blob==12.22.0', # this is for blob exporter
  'boto3==1.34.131', # this is for aws exporter
  'llama-index-vector-stores-opensearch==0.1.10',
  'haystack-ai==2.3.0',
  'llama-index-llms-azure-openai==0.1.9',
  'requests-aws4auth==1.2.3',
  'opensearch-haystack==1.2.0',
  'langchainhub==0.1.21',
  'chromadb==0.4.22',
  'flask',
  'opentelemetry-instrumentation-flask',
  'langgraph==0.2.68',
  'opendal==0.45.14'
]

azure = [
    'azure-storage-blob==12.22.0',
]

aws = [
    'boto3==1.35.19',
]

[project.urls]
Homepage = "https://github.com/monocle2ai/monocle"
Issues = "https://github.com/monocle2ai/monocle/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/monocle_apptrace"]

[tool.hatch.build.targets.sdist]
ignore-vcs = true

[tool.hatch.build]
exclude = [
  "/.*",
  "*.txt",
  "Pipfile",
  "/data",
  "/docs",
  "/tests",
  "*.yml",
]

[tool.pylint]
max-line-length = 120
disable = [
    "C0103", # (invalid-name)
    "C0114", # (missing-module-docstring)
    "C0115", # (missing-class-docstring)
    "C0116", # (missing-function-docstring)
    "R0903", # (too-few-public-methods)
    "R0913", # (too-many-arguments)
    "W0105", # (pointless-string-statement)
]
